ui/opensnitch_ui.desktop: Drop MimeType declaration.
The desktop entry currently lists support for a mime type, but does not
provide codes like %f, %F, %u or %U for the Exec key.

If the application can indeed handle files of the listed mime types, it
should specify a way to pass the filenames as parameters, or it should
not specify the MimeType key to begin with.