Workaround for crashes parsing .desktop files
We parse .desktop files to get the icon of a program, but sometimes we
can't parse the Name translation due to unicode encoding problems.

Besides, on some distributions there're .desktop files without Exec=
line, so we also crash.

With this workaround we miss (mainly) the icon of a program, but at
least we won't crash.

It should help with #5.