UI: add nodes view
Display the list of nodes with its metadata.
Clicking on the IP of a node displays the stats for that node.

BUG: The original behaviour was to open a new window per node. However,
right now only one gets updated, and the others have no data.
So for now, this feature has been disabled.

Besides, if there's more than one node, only the stats of a node is
displayed on the UI, so you will see a weird blinking.