ui: set an alternative theme if the default one fails
In some scenarios some button icons are not available.

In these cases we'll first try to use the fallback theme, f it's
configured. If the buttons are still not available, list the themes
installed, and determine which one has the buttons we need, then use it.

If all this fails, then display a warning message to the user.

closes #53