ui, rules: replace rules received from the daemon
We added UPSERTS, to update the time of a rule when a connection matched
that rule.

However UPSERTS in SQLite weren't introduced until v3.24.x, thus it
causes errors on older versions (like the ones shipped with Ubuntu
16/18).

On the other hand, we need to replace the rules once we receive them
from the daemon, to reflect on the GUI any change made on the rules by hand.

More info: #344