ui, rules editor: fixed parsing duration
When adding a new rule we were getting the duration as it appeared on
the combo box. It was used to create a Golang time.

However translators were translating the duration strings (30s, 5m,
etc), causing errors when saving the rule.

Now we just look for the index of the selected duration.