Unify debian build rules to build both binary debs from same rule set
Added missing build dependency for qttools5-dev-tools.

Note, the install of bpf module is removed because its build rules are not
currently used.  This should be fixed.

The build fail because the build uses go build rules and not the
makefile, and thus the protocol files are not generated in the correct
location.

Added github CI build test for Debian package.  Note, this build include a
workaround for the lack of upstream tarball.  Might be better to create
a tarball from the current source directory first, but then debian/changelog
need to be updated with a version entry for this tarball too.