ui: reverting in-memory DB cached file URI
In 5b5e2714aef9eb6752d15f41c23a69b013764f1e we added support for
in-memory cached DB, to allow delete old events from memory (#844).

Unfortunately, on some systems this URI (file::memory:?cache=shared)
creates a file on disk on user's home.

This file is in the end a DB, so if users want to delete old events,
they'll have to save events to disk.

On the other hand, when in-memory DB is selected, we now disable the
option to delete old events from memory.

Closes #857