rpm pkgs: changed autostart path
Previously after installing the rpm GUI package, we created a symlink to
our .desktop file for all users, under /home/*/.config/autostart/

If the path didn't exist we created it, unfortunately as root, which
caused some problems with other applications like Gnome Tweaks.

Now a link is created under /etc/xdg/autostart/, which is what we use
with the debian packages.

Closes: #1068