added more debug info to the dns ebpf hook
Added the path to the libc as well as the calculated offset for the
uprobe.
Don't return on the first error found loading a uprobe, instead try all
the uprobes and return if the loaded uprobes are 0.

(cherry picked from commit 55678b4d3b19cc9cbfe3cadc7a829ed4ea18330c)