on aarch64 send exec events directly to userspace
On 68c2c8ae1aeb7356f9d90e88e2129465c472e646 we excluded failed execve*
calls from being delivered to userspace, in order to get the binary that
was executed and avoid errors/confusion.

But on aarch64, it seems that we fail to save the exec event to a map,
so the event is never delivered to userspace.

So for the time being, send the exec events as soon as they arrive on
aarch64, without checking if the call failed.

(cherry picked from commit c118058dd8d25d42d36a6ae00414db96fb53e23f)