collect and display bytes sent/recv per process
New feature to collect and display bytes sent/received per process.

 - it only works with 'ebpf' monitor method.
 - the information is collected on kernel space and sent to the daemon:
   - when the connection socket is closed.
   - every 2s on large transfers. On this case the bytes are
     accumulated.

The daemon sends the events to the server (GUI), where the information
is added to the DB.

The information is displayed on the GUI:
 - on the statusbar in real-time (based on the refresh interval defined).
 - on the Applications tab.

By right clicking on the Applications tab headers, the user can reset
Tx/Rx stats, and grouping bytes per unit (default) or not.

Finally, the rx/tx stats are deleted based on the preferences options.