ui,prefs: allow to configure server (GUI) address
Up until now it when changing the node address, you changed the server
adress as well (i.e: the address where the GUI is listening for
connections).

This simplified configuration when you had 1 node and addresses were
changed at the same time. But in other situations, you were not able to
change the server (GUI) address from the Preferences dialog.

So now:
 - the address of each node is changed from the Nodes tab.
 - the address of the server (GUI) is changed from the UI tab.