eBPF: fixed obtaining connection details
Sometimes we failed to retrieve the details of a connection, which
inadvertently caused us to block certain applications (Parsec).

By using inet_dgram_connect() we're now able to identify the PID of
a connection in more cases, such as in #1246.

Note: it's excluded from armhf, since I've been unable to make it work
on that platform.