make connections flushing configurable
By default when adding the interception rules, we were killing all
existing connections, to force them go to the netfilter queue.

However in some environments this is not acceptable, so now it's configurable.

Besides, we were doing this only for nftables, so now it also works for
iptables.

This option will only take effect if FlushConnsOnStart is true, and the
user has (re)configured the ProcMonitorMethod option (typically on
start, or when changing it manually).

cherry picked from commits:
  8e9c1d2178410854e8410e9cea7df6ceae229160
  8935bfe6b04fb5f2748adee7f1853ab3bd5b1963

related commits:
  661e3da48f0cba584756002a9d6b5d0eea1cb11c
  1e6d2d464b4663cc73b55c470efe0a3b2eb982c3