ui: use HighContrast icons for dark qt-material themes
By default we use the system icons theme in the GUI.
The Desktop Environment usually is responsible of switching between
dark and light styles.

However if the user selected a qt-material dark theme from the
Preferences, the icons were not updated, and sometimes they were
practically invisible.

Now if the user selects a qt-material dark theme, we set the
HighContrast icon theme, which ensures that all icons are visible
against dark backgrounds.